Добавил получение условий предоставления и регистрацию займа

+++ app/src/main/java/com/test/a2021_q4_tyukavkin/data/model/LoanConditionsDTO.kt
@@ -1,9 +1,7 @@
 package com.test.a2021_q4_tyukavkin.data.model
 
-import java.math.BigDecimal
-
 data class LoanConditionsDTO (
-    val maxAmount: BigDecimal, //TODO Уточнить тип number
+    val maxAmount: Float, //TODO Уточнить тип number
     val percent: Double,
     val period: Int
 )
